rating fix: help UMS to update cached element to update rating info

+++ backend/nextcp2-runtime/src/main/java/nextcp/rest/DtoBuilder.java
@@ -360,7 +360,7 @@ public class DtoBuilder
                             String strFileId = n.getTextContent();
                             if (NumberUtils.isParsable(strFileId))
                             {
-                                ids.umsAudiotrackId = Integer.parseInt(strFileId);
+                                ids.umsAudiotrackId = Long.parseLong(strFileId);
                                 log.debug("audiotrackId : " + strFileId);
                             }
                         }
